104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -34,12 +34,12 @@ return (length);
 int main(void)
 {
 int count, initial;
-unsigned int f1 = 1, f2 = 2, sum, max = 100000000, f1o = 0, f2o = 0, sumo = 0;
+unsigned long f1 = 1, f2 = 2, sum, max = 100000000, f1o = 0, f2o = 0, sumo = 0;
 
 for (count = 1; count <= 98; count++)
 {
 	if (f1o > 0)
-		printf("%u", f1o);
+		printf("%lu", f1o);
 	initial = numlength(max) - 1 - numlength(f1);
 
 	while (f1o > 0 && initial > 0)
@@ -48,7 +48,7 @@ for (count = 1; count <= 98; count++)
 	initial--;
 	}
 
-	printf("%u", f1);
+	printf("%lu", f1);
 
 	sum = (f1 + f2) % max;
 	sumo = f1o + f2o + (f1 + f2) / max;
